requirements.txt
databricks-vector-search
transformers
pandas
requests
numpy
